test/e2e/multik8s/replication: separate suite_test.go into 3 files
This commit separates the tests in the test/e2e/multik8s/replication/suite_test.go
into 3 different files, i.e., full_backup_test.go, incr_backup_test.go,
and misc_test.go. The tests in each file has each different label
("full-backup", "incr-backup", or "misc"), and these labels are
used to parallelize the tests in the CI.

There should be no functional changes in this commit.
